feat: remove posts, add offers components
