+++ components/offers/offer-detail.tsx
@@ -1,46 +1,25 @@
 'use client';
 
 import { Dialog, DialogContent, IconButton } from '@mui/material';
-import { Post } from 'prisma';
 import dayjs from 'dayjs';
 import { ReactMarkdown } from 'react-markdown/lib/react-markdown';
 import remarkGfm from 'remark-gfm';
 import { XMarkIcon } from '@heroicons/react/24/outline';
-import { getPostTitle, getPrice } from '@/lib/post';
+import { OfferResponse } from '@/generated/graphql';
+import { getPrice } from '@/lib/formatter';
 import UserProfile from '../users/user-profile';
 import ImageSlider from '../base/image-slider';
 
-export default function PostDetailCard({
-  post,
+export default function OfferDetail({
+  offer,
   open,
   handleOpen,
 }: {
-  post: Post;
+  offer: OfferResponse;
   open: boolean;
   handleOpen: React.MouseEventHandler;
 }) {
   const sizes = 'w-full h-full';
-
-  const convertPrice = (price: string): string => {
-    const numReg = /(\d+)/;
-    const matchReg = price.match(numReg);
-
-    if (!matchReg) {
-      return 'invalid ipt';
-    }
-    const amount = parseInt(matchReg[0], 10);
-
-    if (price.includes('천원')) {
-      return `${new Intl.NumberFormat('ko-KR').format(amount)}원`;
-    }
-    if (price.includes('만원')) {
-      return `${new Intl.NumberFormat('ko-KR').format(amount * 10000)}원`;
-    }
-    if (price.includes('백만원')) {
-      return `${new Intl.NumberFormat('ko-KR').format(amount * 1000000)}원`;
-    }
-    return `${new Intl.NumberFormat('ko-KR').format(amount)}원`;
-  };
 
   return (
     <Dialog
@@ -86,7 +65,7 @@ export default function PostDetailCard({
         <div className="md:flex md:flex-row justify-center">
           <div className="rounded-tl-md rounded-bl-none rounded-tr-md md:rounded-tl-md md:rounded-bl-md md:rounded-tr-none md:w-[50%]">
             <ImageSlider
-              images={post.images}
+              images={offer.images}
               sizes={sizes}
               width={760}
               height={760}
@@ -97,30 +76,30 @@ export default function PostDetailCard({
               <div className="border-b-[1px]">
                 <div className="flex flex-row gap-2 text-sm md:text-base items-center pl-2">
                   <UserProfile
-                    user={post.user}
+                    user={offer.seller}
                     displayAvatar
                     displayUsername
                     displayDM
                   />
                   <div className="justify-self-end text-[10px] md:text-xs text-gray-600">
-                    {dayjs(post.createdAt).fromNow()}
+                    {dayjs(offer.createdAt).fromNow()}
                   </div>
                 </div>
                 <div className="p-2 flex flex-col gap-2 mt-1">
-                  <div className="text-[18px] font-normal">
-                    {getPostTitle(post)}
-                  </div>
+                  <div className="text-[18px] font-normal">{offer.name}</div>
                   <div className="flex md:text-base justify-self-end  mt-[4px] items-center mb-[16px]">
                     <span className="text-[18px] font-bold">
-                      {convertPrice(getPrice(post))}
+                      {getPrice(offer.price)}
                     </span>
                   </div>
                 </div>
               </div>
               <div className="p-2 pt-4 text-sm md:text-base md:h-[30rem] overflow-y-auto">
-                <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                  {post.content}
-                </ReactMarkdown>
+                {offer.description && (
+                  <ReactMarkdown remarkPlugins={[remarkGfm]}>
+                    {offer.description}
+                  </ReactMarkdown>
+                )}
               </div>
             </div>
           </div>
